Move GoogleIdentityProvider to separate package.
